Factor out s-expression parser
